- add new methods NativeUi

+++ src/java/Boot/src/com/core/api/engine/ui/types/Element.java
@@ -1,9 +1,9 @@
 package com.core.api.engine.ui.types;
 
 public class Element {
-    public int x, y;
+    public float x, y;
     public String material;
-    public Element(int x, int y, String material){
+    public Element(float x, float y, String material){
         this.x = x;
         this.y = y;
         this.material = material;
@@ -14,10 +14,10 @@ public class Element {
     String getType(){
         return "null";
     }
-    public int getX(){
+    public float getX(){
         return x;
     }
-    public int getY(){
+    public float getY(){
         return y;
     }
 }
